Added log
When I use the keyidentifier returned from getKeySSI -> I can load DSU and read normaly 

When I use sReadSSI derived from seedSSI -> I can load a DSU but it is not the same. The file I created before is not there. That's why I get  an error "UnhandledPromiseRejectionWarning: TypeError: Cannot read property 'toString' of undefined" when I try to read the data